+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/events/ClientEvents.java
@@ -6,6 +6,7 @@ import net.mehvahdjukaar.supplementaries.client.renderers.CapturedMobCache;
 import net.mehvahdjukaar.supplementaries.client.screens.ConfigButton;
 import net.mehvahdjukaar.supplementaries.common.block.blocks.RopeBlock;
 import net.mehvahdjukaar.supplementaries.common.events.overrides.InteractEventOverrideHandler;
+import net.mehvahdjukaar.supplementaries.common.misc.AntiqueInkHelper;
 import net.mehvahdjukaar.supplementaries.configs.ClientConfigs;
 import net.mehvahdjukaar.supplementaries.integration.CompatHandler;
 import net.mehvahdjukaar.supplementaries.integration.QuarkCompat;
@@ -49,7 +50,6 @@ public class ClientEvents {
                     tc.getKey().equals("item.durability")).findFirst();
             r.ifPresent(components::remove);
         }
-
     }
 
     @EventCalled