+++ changelog.md
@@ -1,1 +1,7 @@
-* fixed an issue with supp squared
+* Antique ink can be applied to written books turning them into Tattered books (thanks to artkek for gui texture)
+* Tattered books have unique textures and cant be duplicated. They also use the antiquable font
+* Weathered maps have a unique texture too
+* Written and Writable books are now placed vertically instead of horizontally
+* You can now pick up a placed book by shift clicking on it
+* Fixed a bug where crossbows rope arrows wouldnt have their custom texture
+* Fixed some JEI recipes